[test] move tests module to tests/it
